Updated chess_app.csv to include a different set of games.
This csv should be considered the definitive version. It contians data on 500k games taken randomly from a pgn form lichess database (from 2017, due to size reasons).  Because its slightly over 100MB in size, github doesnt let it be committed, and so I include a .7z file which you should extract and then keep in your folder without pushing to github.
Functionality for chess_app.py does not change, and it should still run quickly enough to not be cumbersome.
Bc of this, I remove the kaggle dataset and notebook, which only contain the ptrograms to generate a preprocessed .csv from the kaggle dataset.